feat: Add StudentsCourse page to Blazor project
Add the StudentsCourse page to the Blazor project with a code-behind class. The code-behind class shows the properties necessary to really interact with the API: an IHttpClientFactory, an IConfiguration (or other way to get the URL of the API), and how to get and parse the request.

I had issues getting the deserialization of the Course object to work. This required a tweak in the API project so that entities are serialized using their exact capitalization.